refactor: deleteEntry api 분리
