+++ src/apis/premieres/postEntry.ts
@@ -1,14 +1,9 @@
 import { axiosInstance } from '../axiosInstance';
 import { END_POINTS } from '../../constants/api';
-import { PaymentsFormData, PaymentsFormDataResponse, EntryFormData } from '../../types/premieres';
+import { PaymentsFormData, PaymentsFormDataResponse } from '../../types/premieres';
 
 interface EntryParams extends PaymentsFormData {
     premiereId: string;
-}
-
-interface EntryCancelParams extends EntryFormData {
-    premiereId: string;
-    orderId: string;
 }
 
 export const postEntry = async ({ premiereId, ...information  } : EntryParams): Promise<PaymentsFormDataResponse> => {
@@ -19,12 +14,3 @@ export const postEntry = async ({ premiereId, ...information  } : EntryParams): 
   return response.data;
 };
 
-export const deleteCancelEntry = async ({ premiereId, ...information  } : EntryCancelParams): Promise<PaymentsFormDataResponse> => {
-  const response = await axiosInstance.delete(
-    END_POINTS.PREMIERE_CANCEL_ENTRY(premiereId),
-    {
-      data: information,
-    }
-  );
-  return response.data;
-};
